Move ElectricLogic SignalElectric stuff around
