+++ src/faebryk/library/PowerSwitchStatic.py
@@ -18,5 +18,5 @@ class PowerSwitchStatic(F.PowerSwitch):
 
     def __preinit__(self):
         self.power_in.connect(self.switched_power_out)
-        self.logic_in.connect_reference(self.power_in)
+        self.logic_in.reference.connect(self.power_in)
         self.logic_in.set(True)
